requirements.txt
absl-py==0.5.0
astor==0.7.1
cycler==0.10.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.0
numpy==1.15.2
pkg-resources==0.0.0
protobuf==3.6.1
pyparsing==2.2.2
python-dateutil==2.7.3
PyYAML==3.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.11.0
tensorflow-gpu==1.11.0
termcolor==1.1.0
tqdm==4.28.1
Werkzeug==0.14.1
